// backend/src/statistics/dto/statistics-update.dto.ts
export class StatisticsUpdateDto {
  number_visits?: number; // количество посещений
  clicks_voice_button?: number; // количество нажатий на кнопку озвучить
  number_voiced_characters?: number; // количество озвученных символов
  number_payments?: number; // количество оплат
  amount_payments?: number; // сумма оплат
  number_repeated_payments?: number; // количество повторных оплат
}